miri: reduce code duplication in SSE/SSE2 bin_op_* functions
